requirements.txt
# SOC_Intelligence - Unified Requirements

# =============================================================================
# Core Framework
# =============================================================================
streamlit>=1.28.0,<2.0.0
fastapi>=0.104.0  # NEW - For API layer
uvicorn>=0.24.0   # NEW - ASGI server

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.1.0
polars>=0.19.0
pyarrow>=14.0.1
numpy>=1.26.0

# =============================================================================
# Database
# =============================================================================
duckdb>=0.9.0,<1.0.0
psycopg2-binary>=2.9.9
impyla>=0.18.0
thrift>=0.16.0
thrift-sasl>=0.4.3
pure-sasl>=0.6.2

# =============================================================================
# Data Validation & Serialization
# =============================================================================
pydantic>=2.5.0
jsonschema>=4.20.0
orjson>=3.9.10

# =============================================================================
# Network Analysis
# =============================================================================
pyshark>=0.6.0
scapy>=2.5.0

# =============================================================================
# Visualization
# =============================================================================
plotly>=5.17.0
altair>=5.1.0

# =============================================================================
# Template Rendering & Reports
# =============================================================================
jinja2>=3.1.2
markupsafe>=2.1.3

# =============================================================================
# AI/LLM (Optional - CPU Only)
# =============================================================================
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.1
tokenizers>=0.15.0
sentencepiece>=0.1.99
huggingface-hub>=0.20.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
regex>=2.8.2
click>=8.1.7
requests>=2.31.0
psutil>=5.9.6

# =============================================================================
# Logging
# =============================================================================
coloredlogs>=15.0

# =============================================================================
# Django (for SecAI PostgreSQL integration)
# =============================================================================
Django>=4.2.0

# =============================================================================
# Machine Learning
# =============================================================================
scikit-learn>=1.3.0

# =============================================================================
# Production Server
# =============================================================================
gunicorn>=21.2.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# =============================================================================
# Development (optional)
# =============================================================================
black>=23.12.0
isort>=5.13.0
mypy>=1.7.0
flake8>=6.1.0
